re organize txs
